Extracted Accordion into different components - Extracted AskQuestion
