+++ src/components/PDP/AskQuestionOpenTrigger.tsx
@@ -1,0 +1,9 @@
+import { Button } from '../ui/button';
+
+export default function AskQuestionOpenTrigger() {
+  return (
+    <Button className="mx-auto my-10 mt-9 flex h-12 w-[216px] rounded border border-[#1A1A1A] bg-[#1A1A1A] text-lg font-bold  uppercase text-white hover:bg-transparent hover:text-[#1A1A1A]">
+      ask a question
+    </Button>
+  );
+}
